grt: address comments

+++ src/grt/src/fastroute/include/FastRoute.h
@@ -125,7 +125,7 @@ class FastRouteCore
                      int x2,
                      int y2,
                      int layer,
-                     int reducedCap,
+                     uint16_t reducedCap,
                      bool isReduce);
   void applyVerticalAdjustments(const odb::Point& first_tile,
                                 const odb::Point& last_tile,
